requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
langchain==0.2.1
langchain-community==0.2.1
langchain-text-splitters>=0.2.0,<0.3.0
chromadb==0.4.18
pypdf==3.17.4
python-docx==1.1.0
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.7.1
redis==6.2.0
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
huggingface_hub==0.20.3


